requirements.txt
numpy
matplotlib
pandas
scikit-learn
pillow
jupyter
tensorflow
streamlit
streamlit-drawable-canvas
